lib: bh_arc: Move init steps to lib_bharc
Move initialization steps into lib_bharc SYS_INIT and
out of SMC main.
