+++ lib/tenstorrent/bh_arc/CMakeLists.txt
@@ -8,6 +8,10 @@
 zephyr_library_named(bh_arc_recovery)
 zephyr_library_include_directories_ifdef(CONFIG_TT_BH_ARC ${ZEPHYR_CURRENT_MODULE_DIR}/include)
 
+if(EXISTS ${APPLICATION_SOURCE_DIR}/VERSION)
+  zephyr_library_compile_definitions(HAS_APP_VERSION=1)
+endif()
+
 zephyr_library_sources(
 # zephyr-keep-sorted-start
   asic_state.c
@@ -16,6 +20,7 @@ zephyr_library_sources(
   cm2dm_msg.c
   dw_apb_i2c.c
   harvesting.c
+  init.c
   led.c
   log.c
   msgqueue.c
